Add files via upload
I learned how to swap two numbers. I usually make a variable as temp and store value in the temp and swap two numbers. However, in Python, there is easy way to swap two numbers. Like a,b = b,a. Also, I learned pop function. If I type the code as a.pop(), then it will remove last input of a list. If I want to remove first index of list then I can write the code like a.pop(0).