Add files via upload
This is the problem combine two list in ascending order. We can user sorted() method. However, the time complexity of sorted is nlogn. If we use for loop or while loop one time time complexity will decrease to n. So I used while loop in here to decrease time complexity